fix: add_header tests
 - add Sync for test `FromRequestParts`
 - move TestRequest header tests into test_request.rs
 - add test for `clear_headers`
